drivers/rawexec: PR comments and feedback
