+++ drivers/rawexec/driver_unix_test.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-package raw_exec
+package rawexec
 
 import (
 	"context"
